% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Function to calculate pairwise distances from DNA sequences}
\usage{
calculate_distance(dnaset, min_eligible_count = 2)
}
\arguments{
\item{dnaset.}{DNAStringSet object}

\item{sequence_type.}{outgrowth or provirus (case-insensitive)
Used to determine error thresholds when calculating diversity metrics WFPS and WFPScodons}

\item{min_eligible_count.}{Minimum number of eligible sequences to accept}
}
\value{
Returns a list of input parameters and result of different diversity metrics.
Currently supported diversity metrics include:
\itemize{
\item mean pairwise distance from the "raw" model (rawMPD)
\item mean pairwise distance from the "TN93" model (tn93MPD)
\item nucleotide diversity from the "raw" model (rawPI)
\item nucleotide diversity from the "TN93" model (tn93PI)
\item weighted fraction of polymorphic sites (WFPS)
\item WFPS at third codon positions only (WFPScodons)
}
}
\description{
Function to calculate pairwise distances from DNA sequences
}
\examples{
dist <- calculate_distance(dnaset = mydnaset, min_eligible_count = 2)

}
