% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_alignment_width.R
\name{check_alignment_width}
\alias{check_alignment_width}
\title{Function checks if the alignment width is at least the required minimum}
\usage{
check_alignment_width(dnaset, min_alignment_width = 9)
}
\arguments{
\item{dnaset.}{DNAStringSet object}

\item{min_alignment_width.}{Minimum alignment width to accept}
}
\value{
Returns TRUE or FALSE
}
\description{
Function checks if the alignment width is at least the required minimum
}
\examples{
pass1 = good_alignment_width(dnaset = trimmed, min_alignment_width = 9)

}
