% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_eligibility.R
\name{check_eligibility}
\alias{check_eligibility}
\title{Function to check eligibility}
\usage{
check_eligibility(dna_bin, min_sequence_count, min_seq_width)
}
\arguments{
\item{dna_bin.}{DNAbin object}

\item{min_sequence_count.}{Minimum sequence count allowed}

\item{min_sequence_width.}{Minimum sequence width allowed}
}
\value{
Returns TRUE if the criteria are passed, otherwise FALSE.
}
\description{
This function counts the number of sequences in the input fasta file (DNAbin object)
checks if they are not less than the specified minimum sequence count
also checks if the alignment width is not less than the specified minimum sequence length
}
\examples{
eligible <- check_eligibility(dna_bin, 2, 9)

}
