% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_eligible_sequences.R
\name{count_eligible_sequences}
\alias{count_eligible_sequences}
\title{This Function ensures the remaining sequence count is at least the required minimum}
\usage{
count_eligible_sequences(dnaset, min_eligible_count = 2)
}
\arguments{
\item{dnaset.}{DNAStringSet object}

\item{min_eligible_count.}{Minimum number of eligible sequences to accept}
}
\value{
Returns TRUE or FALSE
}
\description{
This Function ensures the remaining sequence count is at least the required minimum
}
\examples{
pass1 = good_alignment_width(dnaset = trimmed, min_alignment_width = 9)

}
