% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_at_each_position.R
\name{diversity_at_each_position}
\alias{diversity_at_each_position}
\title{Function to calculate diversity at each position in an alignment}
\usage{
diversity_at_each_position(fasta, errorthreshold, variants)
}
\arguments{
\item{fasta.}{DNAbin object}

\item{errorthreshold.}{Sequencing error threshold. Usually 0.01 for NGS, 0 for prominent outgrowth viruses}

\item{variants.}{expected nucleotide characters.}
}
\value{
Diversity score at each position of the alignment
}
\description{
This function checks the diversity contribution of each variant at each position.
If frequency of minor variants at a position is less than the sequencing error threshold,
that position contributes a diversity score of zero
}
\examples{
div_list <- diversity_at_each_position(dna_bin, 0, c("A","C","G","T","-"))

}
