% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnaStringSet_to_dnabin.R
\name{dnaStringSet_to_dnabin}
\alias{dnaStringSet_to_dnabin}
\title{Function to convert DNAStringSet objects to DNAbin}
\usage{
dnaStringSet_to_dnabin(dnaset)
}
\arguments{
\item{dnaset.}{DNAStringset object}
}
\value{
Returns DNAbin version of DNAStringSet
}
\description{
Function to convert DNAStringSet objects to DNAbin
}
\examples{
dnabin = dnaStringSet_to_dnabin(notgappy)

}
