% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_non_nucleotides.R
\name{filter_non_nucleotides}
\alias{filter_non_nucleotides}
\title{This Function removes sequences containing non-nucleotide characters
beyond a set threshold}
\usage{
filter_non_nucleotides(
  dnaset,
  valid_nucleotides = c("A", "C", "G", "T"),
  min_alignment_width = 9,
  threshold = 0.25
)
}
\arguments{
\item{dnaset.}{DNAStringSet object}

\item{min_alignment_width.}{Minimum required width for dnaset alignment}

\item{valid_nucleotides.}{Acceptable nucleotide characters (upper or lower case)}

\item{threshold.}{Max acceptable proportion of non-nucleotide characters in alignment}
}
\value{
Returns DNAStringSet object, filtered
}
\description{
This Function removes sequences containing non-nucleotide characters
beyond a set threshold
}
\examples{
notgappy = filter_sequences_by_non_nucleotides(dnaset = trimmed, threshold = 0.25)

}
