% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_overall_non_gap_codons.R
\name{locate_overall_non_gap_codons}
\alias{locate_overall_non_gap_codons}
\title{Function to locate the portion of an alignment between the first and last non-gap codons,
in any of the sequences in the alignment.}
\usage{
locate_overall_non_gap_codons(dna_bin)
}
\arguments{
\item{dna_bin.}{DNAbin object}
}
\value{
index1,index2.
}
\description{
Function to locate the portion of an alignment between the first and last non-gap codons,
in any of the sequences in the alignment.
}
\examples{
indices = locate_overall_non_gap_codons(dna_bin)
first = indices[1]
last = indices[2]

}
