% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_TSI.R
\name{predict_TSI}
\alias{predict_TSI}
\title{Function to predict time since infection given a distance}
\usage{
predict_TSI(hiv_region, distance_metric, x_new)
}
\arguments{
\item{hiv_region.}{Character. Currently supported HIV hiv_regions (selected due to better performance) include:
\itemize{
\item Envgp120V5
\item Polp51RT
\item Gagp17Matrix
\item Envgp41
}}

\item{distance_metric.}{Numeric. Currently supported distance metrics include:
\itemize{
\item mean pairwise distance from the "raw" model (rawMPD)
\item mean pairwise distance from the "TN93" model (tn93MPD)
\item nucleotide diversity from the "raw" model (rawPI)
\item nucleotide diversity from the "TN93" model (tn93PI)
\item weighted fraction of polymorphic sites (WFPS)
\item WFPS at third codon positions only (WFPS_codons)
}}

\item{x_new.}{Numeric. distance to be used to estimate time since infection.}
}
\value{
predicted, lower bound of 95\% credible interval, upper bound of 95\% credible interval.
}
\description{
This function selects markov chain monte carlo posterior samples,
for the specified HIV region and distance metric.
}
\examples{
TSI_rawMPD = predict_TSI("Gagp17Matrix", "rawMPD", 0.003)
TSI_WFPS = predict_TSI("Gagp17Matrix", "WFPS", 0.003)

}
