% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_viremic_time.R
\name{predict_viremic_time}
\alias{predict_viremic_time}
\title{#' This Function predicts viremic time (and credible interval)
for multiple weighting methods and diversity metrics.}
\usage{
predict_viremic_time(
  distances,
  sequence_type,
  hiv_region = c("Matrix", "RT", "gp41", "gp41_and_RT_Mean"),
  weights_type = "None",
  diversity_metrics = c("rawMPD", "tn93MPD", "rawPI", "tn93PI")
)
}
\arguments{
\item{distances.}{List. Result from the function calculate_distance.}

\item{hiv_region.}{Character. Currently supported HIV hiv_regions include:
\itemize{
\item Matrix (especially p17 portion of Matrix in gag)
\item RT (reverse transcriptase region of pol)
\item gp41 (gp41 region of env)
}}

\item{weights_type.}{Character. Choose which weights to use for estimation.
Currently supported weights include:
\itemize{
\item None
\item UniqueseqsAsis
\item UniqueseqsLogTransformed (Recommended)
With default settings, result contain estimates for each weighting method.
}}

\item{diversity_metrics.}{Character.
Choose which diversity metrics to use for estimation
Currently supported diversity metrics include:
\itemize{
\item mean pairwise distance from the "raw" model (rawMPD)
\item mean pairwise distance from the "TN93" model (tn93MPD)
\item nucleotide diversity from the "raw" model (rawPI)
\item nucleotide diversity from the "TN93" model (tn93PI)
\item weighted fraction of polymorphic sites (WFPS)
\item WFPS at third codon positions only (WFPScodons)
By default, result contain estimates for each metric.
}}
}
\value{
Returns dataframe of input specifications
and viremic time estimates for each specification
}
\description{
Based on imput, the function selects markov chain monte carlo posterior samples to use,
for the specified HIV region, weighting method, and diversity metric.
}
\examples{
viremic_time <- predict_viremic_time(distances = dist_list, hiv_region = "gp41")

}
