% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_viremic_time.R
\name{predict_y}
\alias{predict_y}
\title{Thus Function uses posterior samples from a markov chain monte carlo (mcmc)
estimation to predict a viremic time and credible interval
given a single diversity metric and specification of weights
based on the count of eligible unique sequences}
\usage{
predict_y(mcmc_samples, x_new, weight_new)
}
\arguments{
\item{mcmc_samples.}{mcmc list}

\item{x_new.}{Numeric. New diversity estimate to use for predicting viremic time.
Currently supported diversity metrics include:
\itemize{
\item mean pairwise distance from the "raw" model (rawMPD)
\item mean pairwise distance from the "TN93" model (tn93MPD)
\item nucleotide diversity from the "raw" model (rawPI)
\item nucleotide diversity from the "TN93" model (tn93PI)
\item weighted fraction of polymorphic sites (WFPS)
\item WFPS at third codon positions only (WFPS_codons)
}}

\item{weight_new.}{Numeric. Weight to use during estimation.
Options include:
1 (None)
uniqueseqs (raw count of unique sequences)
log(uniqueseqs) (log-transformed counts of unique sequences)}
}
\value{
Retuns list of median viremic time,
and lower and upper bounds of credible interval.
}
\description{
Thus Function uses posterior samples from a markov chain monte carlo (mcmc)
estimation to predict a viremic time and credible interval
given a single diversity metric and specification of weights
based on the count of eligible unique sequences
}
\examples{
res <- predict_y(mcmc_samples = mymcmcobject, x_new = mynewx, weight_new = mynewweight)

}
