% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_reference_sequences.R
\name{remove_reference_sequences}
\alias{remove_reference_sequences}
\title{Function to remove reference sequences
with specified reference identifiers in the sequence headers, if any}
\usage{
remove_reference_sequences(dnaset, ref_identifiers = c("Ref", "HXB2"))
}
\arguments{
\item{dnaset.}{DNAStringSet object}

\item{ref_identifiers.}{Vector of pattern(s) in headers that identify reference sequences}
}
\value{
Returns DNAStringSet object
}
\description{
Function to remove reference sequences
with specified reference identifiers in the sequence headers, if any
}
\examples{
noRefs = remove_reference_sequences(dnaset)

}
