% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_terminal_gappy_codons.R
\name{trim_terminal_gappy_codons}
\alias{trim_terminal_gappy_codons}
\title{Function to trim left and right terminal codons containing non-nucleotide characters,
if any.}
\usage{
trim_terminal_gappy_codons(
  dnaset,
  gap_chars = c("-", "n", "N"),
  gap_threshold = 2
)
}
\arguments{
\item{dnaset.}{DNAStringset object

@param gap_chars. Vector of non-nucleotide characters
to trim at terminal ends of alignment (e.g., "-". "n", "N")}

\item{gap_threshold.}{Codons that contain this number of gap_characters
or more will be trimmed from the ends of the alignment}
}
\value{
Returns DNAStringSet object
}
\description{
Function to trim left and right terminal codons containing non-nucleotide characters,
if any.
}
\examples{
trimmed = trim_terminal_gappy_codons(dnaset)

}
