% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_reference_sequences.R
\name{remove_reference_sequences}
\alias{remove_reference_sequences}
\title{Function to remove reference sequences, if any}
\usage{
remove_reference_sequences(fasta_file_path)
}
\arguments{
\item{fasta_file.}{Character. Fasta file path}
}
\value{
Returns filtered sequences as DNAbin object
}
\description{
This function checks for sequences with the pattern "Ref" or "HXB2"
in the sequence header, and removes these sequences
}
\examples{
dna_bin_filtered <- remove_reference_sequences("data/example_GAG_P17.fasta")

}
